# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wickwatch API",
    "version": "0.0.0",
    "description": "REST API of Wickwatch.\n\n**Authentication:** log in with `POST /api/v1/auth/login` (password and TOTP code); the server sets the\n`ww_session` cookie (HttpOnly, SameSite=Strict). All other endpoints except `/healthz` and\n`/api/v1/auth/*` need it. Viewers may call GET endpoints only; state changes need the `admin` role.\nState-changing requests with an `Origin` header from another site are rejected.\n\n**Errors** are JSON `{ \"error\": \"<code>\" }`. Codes of the broker/runtime adapters: `auth_failed`,\n`not_found`, `unsupported`, `invalid_input`, `timeout`, `unavailable`. Others: `unauthenticated`,\n`forbidden`, `forbidden_origin`, `rate_limited`, `internal` and endpoint-specific codes.\n\n**Time** values are ISO 8601 in UTC; money is in the account currency.",
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ww_session"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/challenge-templates": {
      "get": {
        "summary": "Challenge templates from templates/challenges/*.json",
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "profitTargetPct": {
                            "type": "number",
                            "minimum": 0,
                            "maximum": 1000
                          },
                          "dailyLoss": {
                            "type": "object",
                            "required": [
                              "limitPct",
                              "reference",
                              "resetTime",
                              "timezone"
                            ],
                            "properties": {
                              "limitPct": {
                                "type": "number",
                                "exclusiveMinimum": 0,
                                "maximum": 100
                              },
                              "reference": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": [
                                      "balance-or-equity-at-day-start"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "balance-at-day-start"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "equity-at-day-start"
                                    ]
                                  }
                                ]
                              },
                              "resetTime": {
                                "type": "string",
                                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                              },
                              "timezone": {
                                "type": "string",
                                "minLength": 1
                              },
                              "limitBasis": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": [
                                      "initial-balance"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "day-start"
                                    ]
                                  }
                                ]
                              }
                            }
                          },
                          "maxLoss": {
                            "type": "object",
                            "required": [
                              "limitPct",
                              "type"
                            ],
                            "properties": {
                              "limitPct": {
                                "type": "number",
                                "exclusiveMinimum": 0,
                                "maximum": 100
                              },
                              "type": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "enum": [
                                      "static"
                                    ]
                                  },
                                  {
                                    "type": "string",
                                    "enum": [
                                      "trailing"
                                    ]
                                  }
                                ]
                              }
                            }
                          },
                          "minTradingDays": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 1000
                          },
                          "durationDays": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 3650,
                            "nullable": true
                          }
                        }
                      },
                      {
                        "type": "object",
                        "required": [
                          "id",
                          "firm",
                          "program",
                          "phase",
                          "name",
                          "source",
                          "asOf"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^[a-z0-9][a-z0-9-]*$"
                          },
                          "firm": {
                            "type": "string",
                            "minLength": 1
                          },
                          "program": {
                            "type": "string",
                            "minLength": 1
                          },
                          "phase": {
                            "type": "string",
                            "minLength": 1
                          },
                          "name": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "tradingDayDefinition": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string",
                            "minLength": 1
                          },
                          "asOf": {
                            "type": "string",
                            "minLength": 1
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}